实现分库分表,消费消息日志，并执行mysql语句

+++ library-web/src/main/java/com/ice/sub/library/table/entities/BaseUserInfo.java
@@ -1,13 +1,12 @@
 package com.ice.sub.library.table.entities;
 
-import java.math.BigInteger;
 import java.util.Date;
 
 public class BaseUserInfo {
 
   private Integer id;
 
-  private BigInteger userId;
+  private Long userId;
 
   private String account;
 
@@ -60,11 +59,11 @@ public class BaseUserInfo {
     this.id = id;
   }
 
-  public BigInteger getUserId() {
+  public Long getUserId() {
     return userId;
   }
 
-  public void setUserId(BigInteger userId) {
+  public void setUserId(Long userId) {
     this.userId = userId;
   }
